[refactor] MyPageTest 위치 수정 #50
